Locations: do not throw an error for unknown locations
We are getting thousands of "Location has not been found." errors into Sentry. The issue is that it's not really an error (at least not at this level). It's normal state of the API and therefore it's better to return an empty array (which basically mimics the original API - see https://api.skypicker.com/locations?term=UNKNOWN&locale=en-US). It still returns an error for single location query but that's not used anywhere. New field with appropriate error information should be exposed in the query in case someone actually needs this information.

Another argument for this change is this line in the GraphQL specification:

> Every error must contain an entry with the key message with a string description of the error intended for the developer as a guide to understand and correct the error.

But there is nothing to fix, really. Therefore, I am very dubious about these kind of "errors" because this "error" is not fixable in any way. It's just one of the states of the API - expected response.

This change is potentially breaking.

kiwicom-source-id: be1250acfd2c063f9b1940c535c47cc52941a05d