JS: add support for __DEV__ expression
I've been testing something around Babel using this plugin from different project and I decided to keep it in this project (and share with other teams via NPM plugins). It's very useful for situation when you need to write something for development environment only. This change releases new version of our Babel preset as well as new Eslint config so it can be used right away.

kiwicom-source-id: c2f1e7b0441f61a2b171fe65da9b6f1fd44d266b