Eslint: use custom Jest runner to speedup the checks
Eslint is not able to run in parallel so it's slow with many files. Jest, on the other hand, can do it. Moreover, we can take an advantage of our TestsRunner in the future. It also automatically corrects the code if you are not in CI (autofix). Why didn't I use `jest-runner-eslint` library? It's because the output reporter sucks and it also corrupts the output. Speed improvement? From 26s to 13s (50%) and I just started... :)

kiwicom-source-id: 8dd27dfec736ab3ba8ab0651a0205e06d4d61a3d