Run Eslint only on changed files
This is gonna be tricky when we change some rule (we have to run `yarn lint`) but on the other hand Eslint is not such a big deal for production code and it could actually help us a lot when applying new rules across our codebase (changed some file? update the style as well). The biggest benefit is speed in monorepo setup of course.

kiwicom-source-id: 5c254fd9cea467e70499a7ecb8d4d7a42ab7e2db