Eslint config: show warnings for 'import/no-anonymous-default-export' rule
There are still many warnings we have to fix internally but I think it makes sense to enable it (even though it can be a bit annoying). We'll see. We can always revert it back. BTW this is one of the rules where we are diverging from configs like Airbnb on purpose.

Closes: https://github.com/kiwicom/eslint-config-kiwicom/issues/1

kiwicom-source-id: d704fd907d7b9592ec6b0fea9dbe93388e9e426b